Lazily load the DefPathHash->DefIndex map in crate metadata.
Even though the on-disk hash table would allow for just borrowing
the data (and thus making "loading" the data from the memory mapped
file essentially free) it currently seems to be non-trivial to keep a
reference into the memory blob. As a consequence the implementation
has to copy the data. This is not very expensive but it's still
quite a bit of unnecessary work for non-incremental builds where that
data currently is never actually used.